third_party/WebKit/LayoutTests/accessibility/scroll-div-sends-notification-expected.txt

One
Two
Three

This test ensures that scrolling the window sends a notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Got notification on container div
PASS container.scrollTop is 500
PASS successfullyParsed is true

TEST COMPLETE
